+++ packages/admin/client/Router.js
@@ -18,7 +18,7 @@ Router.map(function() {
 		}
 	});
 
-	this.route('panelAdminUsers', {
+	this.route('pageAdminUsers', {
 		path: '/admin/users',
 		template: 'pageAdminUsers',
 		layoutTemplate: 'layoutPage',
@@ -30,7 +30,6 @@ Router.map(function() {
 				this.next();
 		},
 		waitOn: function() {
-			Session.set('showSidebar', true);
 			return Meteor.subscribe('usersByDate');
 		},
 		data: function() {
@@ -44,7 +43,36 @@ Router.map(function() {
 		}
 	});
 
-	this.route('panelAdminPlaces', {
+	this.route('pageAdminUser', {
+		path: '/admin/users/:id',
+		template: 'pageAdminUsers',
+		layoutTemplate: 'layoutPage',
+		loadingTemplate: 'pageLoading',
+		onBeforeAction: function () {
+			if(!K.Admin.isMe())
+				Router.go('root');
+			else
+				this.next();
+		},
+		waitOn: function() {
+			return [
+				Meteor.subscribe('userById', this.params.id),
+				Meteor.subscribe('usersByDate')
+			];
+		},
+		data: function() {
+			if(!this.ready()) return null;
+			var users = K.findUsersByDate().fetch(),
+				userIds = _.pluck(users,'_id');
+			userIds = _.without(userIds, K.Profile.id);
+			return {
+				itemSelected: K.userById(this.params.id),
+				items: _.map(userIds, K.userById)
+			};
+		}
+	});	
+
+	this.route('pageAdminPlaces', {
 		path: '/admin/places',
 		template: 'pageAdminPlaces',
 		layoutTemplate: 'layoutPage',
@@ -56,12 +84,39 @@ Router.map(function() {
 				this.next();
 		},
 		waitOn: function() {
-			Session.set('showSidebar', true);
 			return Meteor.subscribe('placesByDate');
 		},
 		data: function() {
 			if(!this.ready()) return null;
 			return {
+				items: _.map(K.findPlacesByDate().fetch(), function(item) {
+					return K.placeById(item._id);
+				})
+			};
+		}
+	});
+
+	this.route('pageAdminPlace', {
+		path: '/admin/places/:id',
+		template: 'pageAdminPlaces',
+		layoutTemplate: 'layoutPage',
+		loadingTemplate: 'pageLoading',
+		onBeforeAction: function () {
+			if(!K.Admin.isMe())
+				Router.go('root');
+			else
+				this.next();
+		},
+		waitOn: function() {
+			return [
+				Meteor.subscribe('placeById', this.params.id),
+				Meteor.subscribe('placesByDate')
+			];
+		},
+		data: function() {
+			if(!this.ready()) return null;
+			return {
+				itemSelected: K.placeById(this.params.id),
 				items: _.map(K.findPlacesByDate().fetch(), function(item) {
 					return K.placeById(item._id);
 				})