+++ packages/categories/plugin.js
@@ -9,7 +9,7 @@ K.Plugin({
 		},
 		panelPlaceEdit: 'panelPlaceEdit_cats',
 		panelAdmin: 'panelAdmin_cats',
-		pageAdminPlace: 'panelPlaceEdit_cats',
+		//pageAdminPlace: 'panelPlaceEdit_cats',
 		pageAdminUser: 'pageAdminUser_cats',
 	},
 	settings: {